copy bootstrap file to http server docroot directory only if httpd service is running

+++ 3_ocp4_vsphere_upi_gen_ign_configs.sh
@@ -4,8 +4,13 @@ source 0_ocp4_vsphere_upi_init_vars
 
 openshift-install --dir $CLUSTER create ignition-configs
 
-[ -f /var/www/html/bootstrap.ign ] && rm -f /var/www/html/bootstrap.ign
-pushd $CLUSTER
-cp bootstrap.ign /var/www/html
-popd
+systemctl status httpd >/dev/null 2>&1
+if [ $? -ne 0 ]; then
+  echo "WARNING:  HTTPD service is not running on this system.  The bootstrap.ign file has been generated but will need to manually be copied to the HTTPD serve
+r." >&2
+  exit 1
+else
+  [ -f /var/www/html/bootstrap.ign ] && rm -f /var/www/html/bootstrap.ign
+  cp ${CLUSTER}/bootstrap.ign /var/www/html
+fi
